# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"NestJs Demo","description":"description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"age":{"type":"number"}},"required":["email","password","role","age"]},"User":{"type":"object","properties":{}},"CreatePermissionDto":{"type":"object","properties":{"name":{"type":"string"},"role":{"type":"string"}},"required":["name","role"]},"CreateShopDto":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"category":{"type":"string"}},"required":["name","address","category"]},"CreateShopItemDto":{"type":"object","properties":{"name":{"type":"string"},"brand":{"type":"string"},"mrp":{"type":"number"},"shopId":{"type":"string"}},"required":["name","brand","mrp","shopId"]}}},"paths":{"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/auth/verifyToken":{"get":{"operationId":"AuthController_verifyToken","parameters":[],"responses":{"200":{"description":"Verify Token"}},"security":[{"JWT":[]}],"tags":["auth"]}},"/api/v1/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"Create User"},"201":{"description":""}}},"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"200":{"description":""}},"security":[{"JWT":[]}],"tags":["users"]}},"/api/v1/users/{id}":{"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete current user"}},"security":[{"JWT":[]}],"tags":["users"]},"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Update current user"},"401":{"description":""}},"security":[{"JWT":[]}],"tags":["users"]},"get":{"operationId":"UserController_logoutUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Log  x out current user"},"401":{"description":""}},"security":[{"JWT":[]}],"tags":["users"]}},"/api/v1/users/permission":{"post":{"operationId":"UserController_createPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePermissionDto"}}}},"responses":{"201":{"description":""}},"security":[{"JWT":[]}],"tags":["users"]}},"/api/v1/shop":{"post":{"operationId":"ShopController_createShop","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateShopDto"}}}},"responses":{"200":{"description":"Create Shop"},"201":{"description":""}},"security":[{"JWT":[]}],"tags":["shop"]},"get":{"operationId":"ShopController_getShops","parameters":[],"responses":{"200":{"description":""}},"security":[{"JWT":[]}],"tags":["shop"]}},"/api/v1/shop/{id}":{"delete":{"operationId":"ShopController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete current Shop"}},"security":[{"JWT":[]}],"tags":["shop"]},"put":{"operationId":"ShopController_updateShop","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Update current shop"},"401":{"description":""}},"security":[{"JWT":[]}],"tags":["shop"]}},"/api/v1/shopItems":{"post":{"operationId":"ShopItemController_createShop","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateShopItemDto"}}}},"responses":{"200":{"description":"Create Shop Items"},"201":{"description":""}},"security":[{"JWT":[]}],"tags":["shopItem"]}},"/api/v1/shopItems/{id}":{"delete":{"operationId":"ShopItemController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete current Shop Items"}},"security":[{"JWT":[]}],"tags":["shopItem"]},"put":{"operationId":"ShopItemController_updateShop","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Update current shop Items"},"401":{"description":""}},"security":[{"JWT":[]}],"tags":["shopItem"]}},"/api/v1/shopItems/{shopId}/items":{"get":{"operationId":"ShopItemController_getShops","parameters":[{"name":"shopId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"get all shop Items with particular shop"}},"security":[{"JWT":[]}],"tags":["shopItem"]}},"/api/v1/google":{"get":{"operationId":"GoogleAuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/google/redirect":{"get":{"operationId":"GoogleAuthController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}}}}}}